+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 Changes:
 * Add button to scroll down when chat is scrolled up(#56)
+* Message timestamps(#55)